requirements.txt
streamlit
pandas
requests
plotly
tensorflow
Pillow
numpy
